+++ app/src/main/java/xyz/goodistory/yumemiassignment/http/Contributor.kt
@@ -1,4 +1,4 @@
-package xyz.goodistory.yumemiassignment
+package xyz.goodistory.yumemiassignment.http
 
 data class Contributor(
     val login: String,
